Calculate shortest path to produce a layout.
The current algorithm prevents a line to be too short or too long.  The
cost for each edge is decided by a simple one dimensional algebraic
function.  I have replaced the simple layout with the new graph layout
in the flashcard screen.  Also the debug screen for the graph shows the
nodes and edges selected by the algorithm highlighted in red.